+++ R/dna.n.q.fun.r
@@ -19,8 +19,7 @@
 #' DNA_N_q_fun(293:400)
 DNA_N_q_fun <- function(w.length) {
   wl.within <- w.length >= 256 & w.length <= 364
-  spectral_weights <- numeric(length(w.length))
-  spectral_weights[w.length < 256] <- NA # the value at 256 nm
+  spectral_weights <- rep(NA_real_, length(w.length))
   spectral_weights[w.length > 364] <- 0.0
   if (any(wl.within)) {
     # avoids error in spline when xout is empty