+++ data-raw/make.data.R
@@ -40,5 +40,6 @@ CIE1951_scotopic_lef.spct <- mutate(CIE1951_scotopic_lef.spct,
                            s.e.response = s.q.response * w.length / 504)
 setResponseSpct(CIE1951_scotopic_lef.spct)
 
-save(SetlowTUV.spct, CIE2008_lef2deg.spct, CIE1924_lef.spct, CIE1951_scotopic_lef.spct,
-     file = "./data/BSWF.spectra.rda")
+save(SetlowTUV.spct, CIE2008_lef2deg.spct, CIE1924_lef.spct,
+     CIE1951_scotopic_lef.spct,
+     file = "./data/BSWF-spectra.rda")